requirements.txt
descope
python-dotenv 